chore: added current state
