+++ bouncer/lp_bot/swap_simulator.ts
@@ -1,17 +1,16 @@
-#!/usr/bin/env -S pnpm tsx
-
 import { newAddress, sleep } from '../shared/utils';
 import { requestNewSwap } from '../shared/perform_swap';
 import { FillOrKillParamsX128 } from '../shared/new_swap';
-
 import { globalLogger as logger } from '../shared/utils/logger';
-import { sendBtc } from '../shared/send_btc';
 import { sendErc20 } from '../shared/send_erc20';
 import { getContractAddress } from '../shared/utils';
 
-
-
-async function doSwap() {
+/**
+ * Performs a swap.
+ * 
+ * @param amountToSwap - The amount of USDT to swap.
+ */
+async function doSwap(amountToSwap: string) {
     const destinationAddressForUsdc = await newAddress('Usdc', 'test');
     const refundAddress = await newAddress('Usdt', 'test');
     const refundParameters: FillOrKillParamsX128 = {
@@ -19,6 +18,7 @@ async function doSwap() {
         refundAddress,
         minPriceX128: '0',
     };
+    logger.info(`Requesting new swap for ${amountToSwap} USDT...`);
 
     const swapParams = await requestNewSwap(
         logger.child({ tag: 'swapSimulator' }),
@@ -32,19 +32,47 @@ async function doSwap() {
     );
 
     const contractAddress = getContractAddress('Ethereum', 'Usdt');
-    await sendErc20(logger, 'Ethereum', swapParams.depositAddress, contractAddress, '100');
+    logger.info(`Sending ${amountToSwap} USDT to ${swapParams.depositAddress}...`);
+
+    await sendErc20(logger, 'Ethereum', swapParams.depositAddress, contractAddress, amountToSwap);
     logger.info(`Usdt sent to ${swapParams.depositAddress}`);
 }
 
-
-async function main() {
+/**
+ * Starts the swap simulator.
+ * 
+ * @param limit - The number of swaps to perform.
+ * @param signal - The signal to abort the swaps.
+ */
+async function startSwapSimulator(limit?: number, signal?: AbortSignal) {
     logger.info(`Lets swap!`);
 
-    for (let i = 0; i < 10; i++) {
-        logger.info(`Swapping ${i} times`);
-        await doSwap();
-        await sleep(2000);
+    if (limit) {
+        logger.info(`Swapping ${limit} times!`);
     }
+
+    let swaps = 0;
+
+    while (true) {
+        let amountToSwap = Math.floor(Math.random() * 1000);
+        await doSwap(amountToSwap.toString());
+        await sleep(2000);
+
+        if (signal) {
+            signal.addEventListener('abort', () => {
+                logger.info('Aborting swap simulator early!');
+                return;
+            });
+        }
+
+        swaps++;
+        if (limit && swaps >= limit) {
+            logger.info('Swapping limit reached!');
+            break;
+        }
+    }
+
+    logger.info('Swap simulator completed!');
 }
 
-main();
+export { startSwapSimulator };
